Move binding related methods into BindingManager.
