+++ polar-core/src/formatting.rs
@@ -177,6 +177,7 @@ pub mod display {
     use crate::rules::Rule;
     use crate::terms::{Operation, Operator, Symbol, Term, Value};
     use crate::vm::*;
+    use crate::bindings::Binding;
 
     impl fmt::Display for Binding {
         fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
