+++ polar-core/src/kb.rs
@@ -4,10 +4,10 @@ use super::counter::Counter;
 use super::rules::*;
 use super::sources::*;
 use super::terms::*;
+pub use super::bindings::Bindings;
 
 /// A map of bindings: variable name → value. The VM uses a stack internally,
 /// but can translate to and from this type.
-pub type Bindings = HashMap<Symbol, Term>;
 
 #[derive(Default)]
 pub struct KnowledgeBase {
