+++ polar-core/src/lib.rs
@@ -2,17 +2,19 @@
 #[macro_use]
 extern crate maplit;
 
-mod debugger;
-pub mod error;
-pub mod formatting;
-mod lexer;
 #[macro_use]
 pub mod macros;
+
+mod bindings;
 mod counter;
+mod debugger;
+pub mod error;
 pub mod events;
 mod folder;
+pub mod formatting;
 mod inverter;
 pub mod kb;
+mod lexer;
 pub mod messages;
 mod numerics;
 pub mod parser;
